package: Add Heroku deploy script
This is used to deploy to Heroku. It's not good practice to use the --force flag blindly. However, in this small app it's acceptable.
It's required as before deployment to Heroku, there is a temporary commit where all required files are moved to root. This causes a conflict when deploying, as the NEW master will be "x commits ahead of OLD master, 1 commit behind".